adding ddp unittest for lstm
