+++ README.md
@@ -108,5 +108,5 @@ cuTAGI is released under the MIT license.
 ```
 
 ```shell
-mpirun -np 2 build/run_tests --gtest_filter=DistributedTest.SimpleCNN_NCCL
+mpirun -np 2 build/run_tests --gtest_filter=DDPOpsTest.*
 ```
